Processing review comments
