+++ services/src/main/resources/META-INF/services/org.keycloak.authentication.AuthenticatorFactory
@@ -53,5 +53,4 @@ org.keycloak.authentication.authenticators.browser.WebAuthnAuthenticatorFactory
 org.keycloak.authentication.authenticators.browser.WebAuthnPasswordlessAuthenticatorFactory
 org.keycloak.authentication.authenticators.access.DenyAccessAuthenticatorFactory
 org.keycloak.authentication.authenticators.access.AllowAccessAuthenticatorFactory
-org.keycloak.authentication.authenticators.sessionlimits.RealmSessionLimitsAuthenticatorFactory
 org.keycloak.authentication.authenticators.sessionlimits.UserSessionLimitsAuthenticatorFactory
